Revert "Added Endpoints and new pages"
