+++ packages/web-app/package.json
@@ -28,7 +28,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1",
-    "react-leaflet-draw": "^0.20.4",
     "react-native-svg": "^15.1.0",
     "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
