+++ packages/web-app/src/pages/HomePage.js
@@ -1,30 +1,51 @@
-import React from "react";
+import React, { useState } from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faThumbsUp, faThumbsDown } from '@fortawesome/free-solid-svg-icons';
+import Leaflet from '../components/leaflet';
+import ImageDisplay from '../components/ImageDisplay';
+import NavigationBar from '../components/NavigationBar';
+import { colors } from '../styles/commonStyles';
 
 const HomePage = () => {
+	const [likeDislike, setLikeDislike] = useState('neither'); // 'like', 'dislike', 'neither'
 
-	// Temporary function to fill the local storage with a user object for testing
-	const fillLocalStorage = () => {
-		localStorage.setItem('user', JSON.stringify({
-							id: '12345689123', // Assuming the user object has an _id property
-							name: 'camilo', // Customize these fields based on your user object
-							email: 'camilo@gmail.com',
-						}));
-	}
+	const handleLike = () => {
+		setLikeDislike(likeDislike !== 'like' ? 'like' : 'neither');
+	};
+
+	const handleDislike = () => {
+		setLikeDislike(likeDislike !== 'dislike' ? 'dislike' : 'neither');
+	};
 
 	return (
-		<div className="flex flex-col items-center justify-center h-screen">
-			<h1 className="text-4xl font-bold">Welcome to GeoGame!</h1>
-			<button className="mt-4 px-4 py-2 font-bold text-white bg-blue-500 hover:bg-blue-700">
-				<a href="/GuessPage">Start Playing</a>
-			</button>
-			<button className="mt-4 px-4 py-2 font-bold text-white bg-blue-500 hover:bg-blue-700">
-				<a href="/CreateTownPage">Create a Town</a>
-			</button>
-			<button className="mt-4 px-4 py-2 font-bold text-white bg-blue-500 hover:bg-blue-700" onClick={fillLocalStorage}>
-				Fill Local Storage
-			</button>
+		<div className="flex min-h-screen w-screen flex-col bg-webPrimary">
+			<NavigationBar />
+			<div className="flex flex-grow flex-col">
+				<div className="flex flex-grow flex-row justify-center">
+					<div className="mb-2 flex w-1/2 flex-col">
+						<ImageDisplay />
+						<div id="thumbs-up" className="mt-3 mr-20 mb-8 flex justify-end space-x-2">
+							<button
+								onClick={handleLike}
+								className={`p-2 rounded ${likeDislike === 'like' ? 'bg-green-500' : 'bg-gray-200'} w-12 h-12 flex items-center justify-center`}
+							>
+								<FontAwesomeIcon icon={faThumbsUp} color={likeDislike === 'like' ? 'white' : 'black'} size="lg" />
+							</button>
+							<button
+								onClick={handleDislike}
+								className={`p-2 rounded ${likeDislike === 'dislike' ? 'bg-red-500' : 'bg-gray-200'} w-12 h-12 flex items-center justify-center`}
+							>
+								<FontAwesomeIcon icon={faThumbsDown} color={likeDislike === 'dislike' ? 'white' : 'black'} size="lg" />
+							</button>
+						</div>
+					</div>
+					<div className="flex w-1/2 flex-grow flex-col pt-4 rounded-4xl">
+						<Leaflet />
+					</div>
+				</div>
+			</div>
 		</div>
 	);
-}
+};
 
 export default HomePage;
