+++ packages/web-app/src/pages/RegisterPage.js
@@ -71,20 +71,12 @@ function RegisterForm() {
 				body: JSON.stringify({ username, email, password }),
 			});
 
-			if (!response.ok) {
-				console.error('Signup failed');
-				return;
-			}
+            if (!response.ok) {
+                throw new Error('Signup failed');
+            }
 
-			const data = await response.json(); // The response from your login/signup route
-			if (data && data.user) {
-				localStorage.setItem('user', JSON.stringify({
-					id: data.user._id, // Assuming the user object has an _id property
-					name: data.user.name, // Customize these fields based on your user object
-					email: data.user.email,
-				}));
-			}
-
+			const data = await response.json();
+			console.log('Signup successful:', data);
 			navigate('/HomePage');
 		} catch (error) {
 			console.error('Error signing up:', error);
